chore: forge fmt

+++ src/EnhancedConsoleLib.sol
@@ -25,7 +25,9 @@ library EnhancedConsoleLib {
     string internal constant BG_YELLOW = "\x1b[43m";
     string internal constant BG_CYAN = "\x1b[46m";
 
-    function log(string memory message) internal pure {
+    function log(
+        string memory message
+    ) internal pure {
         console.log(message);
     }
 
@@ -33,57 +35,86 @@ library EnhancedConsoleLib {
         console.log(string.concat(color, message, RESET));
     }
 
-    function logStyled(string memory message, string memory color, string memory style) internal pure {
+    function logStyled(
+        string memory message,
+        string memory color,
+        string memory style
+    ) internal pure {
         console.log(string.concat(style, color, message, RESET));
     }
 
-    function logRed(string memory message) internal pure {
+    function logRed(
+        string memory message
+    ) internal pure {
         logColored(message, RED);
     }
 
-    function logGreen(string memory message) internal pure {
+    function logGreen(
+        string memory message
+    ) internal pure {
         logColored(message, GREEN);
     }
 
-    function logYellow(string memory message) internal pure {
+    function logYellow(
+        string memory message
+    ) internal pure {
         logColored(message, YELLOW);
     }
 
-    function logBlue(string memory message) internal pure {
+    function logBlue(
+        string memory message
+    ) internal pure {
         logColored(message, BLUE);
     }
 
-    function logMagenta(string memory message) internal pure {
+    function logMagenta(
+        string memory message
+    ) internal pure {
         logColored(message, MAGENTA);
     }
 
-    function logCyan(string memory message) internal pure {
+    function logCyan(
+        string memory message
+    ) internal pure {
         logColored(message, CYAN);
     }
 
-    function logWhite(string memory message) internal pure {
+    function logWhite(
+        string memory message
+    ) internal pure {
         logColored(message, WHITE);
     }
 
-    function logBold(string memory message) internal pure {
+    function logBold(
+        string memory message
+    ) internal pure {
         logStyled(message, WHITE, BOLD);
     }
 
-    function logUnderline(string memory message) internal pure {
+    function logUnderline(
+        string memory message
+    ) internal pure {
         logStyled(message, WHITE, UNDERLINE);
     }
 
-    function logWithBackground(string memory message, string memory color, string memory bgColor) internal pure {
+    function logWithBackground(
+        string memory message,
+        string memory color,
+        string memory bgColor
+    ) internal pure {
         console.log(string.concat(color, bgColor, message, RESET));
     }
 
     function logTable(string[] memory headers, string[][] memory rows) internal pure {
-        require(headers.length > 0 && rows.length > 0 && headers.length == rows[0].length, "Invalid table structure");
+        require(
+            headers.length > 0 && rows.length > 0 && headers.length == rows[0].length,
+            "Invalid table structure"
+        );
 
-        uint[] memory columnWidths = new uint[](headers.length);
-        for (uint i = 0; i < headers.length; i++) {
+        uint256[] memory columnWidths = new uint256[](headers.length);
+        for (uint256 i = 0; i < headers.length; i++) {
             columnWidths[i] = bytes(headers[i]).length;
-            for (uint j = 0; j < rows.length; j++) {
+            for (uint256 j = 0; j < rows.length; j++) {
                 if (bytes(rows[j][i]).length > columnWidths[i]) {
                     columnWidths[i] = bytes(rows[j][i]).length;
                 }
@@ -91,8 +122,8 @@ library EnhancedConsoleLib {
         }
 
         string memory separator = "+";
-        for (uint i = 0; i < headers.length; i++) {
-            for (uint j = 0; j < columnWidths[i] + 2; j++) {
+        for (uint256 i = 0; i < headers.length; i++) {
+            for (uint256 j = 0; j < columnWidths[i] + 2; j++) {
                 separator = string.concat(separator, "-");
             }
             separator = string.concat(separator, "+");
@@ -100,15 +131,15 @@ library EnhancedConsoleLib {
 
         console.log(separator);
         string memory headerRow = "|";
-        for (uint i = 0; i < headers.length; i++) {
+        for (uint256 i = 0; i < headers.length; i++) {
             headerRow = string.concat(headerRow, " ", padRight(headers[i], columnWidths[i]), " |");
         }
         console.log(headerRow);
         console.log(separator);
 
-        for (uint i = 0; i < rows.length; i++) {
+        for (uint256 i = 0; i < rows.length; i++) {
             string memory row = "|";
-            for (uint j = 0; j < headers.length; j++) {
+            for (uint256 j = 0; j < headers.length; j++) {
                 row = string.concat(row, " ", padRight(rows[i][j], columnWidths[j]), " |");
             }
             console.log(row);
@@ -116,11 +147,12 @@ library EnhancedConsoleLib {
         console.log(separator);
     }
 
-
-    function logBanner(string memory message) internal pure {
-        uint length = bytes(message).length;
+    function logBanner(
+        string memory message
+    ) internal pure {
+        uint256 length = bytes(message).length;
         string memory stars = "";
-        for (uint i = 0; i < length + 4; i++) {
+        for (uint256 i = 0; i < length + 4; i++) {
             stars = string.concat(stars, "*");
         }
 
@@ -129,11 +161,11 @@ library EnhancedConsoleLib {
         console.log(string.concat(stars, RESET));
     }
 
-    function padRight(string memory str, uint length) private pure returns (string memory) {
+    function padRight(string memory str, uint256 length) private pure returns (string memory) {
         bytes memory strBytes = bytes(str);
         if (strBytes.length >= length) return str;
         bytes memory result = new bytes(length);
-        uint j = 0;
+        uint256 j = 0;
         for (; j < strBytes.length; j++) {
             result[j] = strBytes[j];
         }
